+++ nomeroff_net/pipes/number_plate_keypoints_detectors/bbox_np_points.py
@@ -82,7 +82,6 @@ class NpPointsCraft(object):
         # load net
         self.net = CRAFT()  # initialize
 
-        print('Loading weights from checkpoint (' + trained_model + ')')
         if is_cuda:
             model = torch.load(trained_model)
             self.net.load_state_dict(copy_state_dict(model))
@@ -100,7 +99,6 @@ class NpPointsCraft(object):
         self.refine_net = None
         if is_refine:
             self.refine_net = RefineNet()
-            print('Loading weights of refiner from checkpoint (' + refiner_model + ')')
             if is_cuda:
                 self.refine_net.load_state_dict(copy_state_dict(torch.load(refiner_model)))
                 self.refine_net = self.refine_net.cuda()
@@ -189,6 +187,7 @@ class NpPointsCraft(object):
                     text_threshold: float = 0.6,
                     link_threshold: float = 0.7,
                     low_text: float = 0.4,
+                    in_zone_only: bool = False,
                     **_) -> Any:
         if quality_profile is None:
             quality_profile = [1, 0, 0, 0]
@@ -241,6 +240,12 @@ class NpPointsCraft(object):
                         points = target_points_variants[idx]
                     else:
                         points = target_points_variants[0]
+                    if in_zone_only:
+                        for i in range(len(points)):
+                            points[i][0] = x0 if points[i][0] < x0 else points[i][0]
+                            points[i][1] = y0 if points[i][1] < y0 else points[i][1]
+                            points[i][0] = x0 + w0 if points[i][0] > x0 + w0 else points[i][0]
+                            points[i][1] = y0 + h0 if points[i][1] > y0 + h0 else points[i][1]
                     all_points.append(points)
                 else:
                     all_points.append([