+++ nomeroff_net/text_detectors/ge.py
@@ -3,6 +3,7 @@ python3 -m nomeroff_net.text_detectors.ge -f nomeroff_net/text_detectors/ge.py
 """
 import torch
 from .base.ocr import OCR
+from nomeroff_net.tools.mcm import get_device_torch
 
 
 class Ge(OCR):
@@ -13,7 +14,6 @@ class Ge(OCR):
         self.max_text_len = 8
         self.max_plate_length = 8
         self.letters_max = len(self.letters)+1
-        self.label_length = 32 - 2
         self.init_label_converter()
 
 
@@ -22,5 +22,7 @@ ge = Ge
 if __name__ == "__main__":
     ocr = Ge()
     ocr.load()
-    y = ocr.predict(torch.rand((1, 3, 50, 200)))
+    device = get_device_torch()
+    xs = torch.rand((1, 3, 50, 200)).to(device)
+    y = ocr.predict(xs)
     print(y)