+++ nomeroff_net/text_detectors/su.py
@@ -3,6 +3,7 @@ python3 -m nomeroff_net.text_detectors.su -f nomeroff_net/text_detectors/su.py
 """
 import torch
 from .base.ocr import OCR
+from nomeroff_net.tools.mcm import get_device_torch
 
 
 class Su(OCR):
@@ -15,7 +16,6 @@ class Su(OCR):
         self.max_text_len = 7
         self.max_plate_length = 7
         self.letters_max = len(self.letters)+1
-        self.label_length = 32 - 2
         self.init_label_converter()
 
 
@@ -24,5 +24,7 @@ su = Su
 if __name__ == "__main__":
     ocr = Su()
     ocr.load()
-    y = ocr.predict(torch.rand((1, 3, 50, 200)))
+    device = get_device_torch()
+    xs = torch.rand((1, 3, 50, 200)).to(device)
+    y = ocr.predict(xs)
     print(y)